requirements.txt
flask 
python-dotenv 
openai
pinecone-client
pinecone